product_pro/requirement.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2024.1
